parser/enrichment/requirements.txt
pymongo>=4.0,<5.0
requests>=2.25,<3.0
python-jose[cryptography]==3.3.0